Add user API docs and testing guide, update controllers and routes
Added API_DOCUMENTATION.md and TESTING_GUIDE.md for user management endpoints and testing instructions. Refactored UserController to merge logic from UserControllerNew, removed UserControllerNew, and restored UserControllerOld for legacy reference. Updated UserRoutes to expose new endpoints for login, user CRUD, password recovery, and email verification.
